# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AURA Forge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeploymentInput {
  networkSelected: string;
  deploymentTxHash: string;
  liveDeployedAddress: string;
}