# lib/api-zod/src/generated/types/ecosystem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AURA Forge API specification
 * OpenAPI spec version: 0.1.0
 */

export type Ecosystem = typeof Ecosystem[keyof typeof Ecosystem];


export const Ecosystem = {
  EVM: 'EVM',
  SOLANA: 'SOLANA',
} as const;